Updated to demonstrate K8S and dotnet app running in K8S cluster.
This was to investigate a DNS SRV possible problem.  Suspicion was
that the name resolution wasn't working properly, and indeed it
does not.

With what exists under the k8s directory, one can set up a K8S
cluster with kind, deploy Couchbase, then build and deploy the
docker image to run the .NET app.